Initial commit
Using a cargo makefile to hopefully adhere more to rust for this project.
I've been thinking about removing the hello.c file and using a dummy c for the SRCS in the regular Makefile.
Its not a nessecary, but its something to look into if I want this to even more rust oriented. Sadly, not everything can be rust here.