Created Cdev wrapper, Renamed Cdev trait to Cdevsw
This wrapper is similar to the uio wrapper for lifetime purposes, we use it so our cdevsw trait doesnt need to deal with pointers. The renaming of the trait is to prevent any conflict & I believe it is a better fit as the functions implemented are used in the cdevsw struct.