Double Boxed dyn Cdevsw to store in si_drv1
This is most likly a terrible way to do this as I made EchoDevice::new() return the Box<Box<dyn Cdevsw>>.
Because of this I need to figure out another solution to my quiesce event.
Oh well, what is learning without trial and errors?

We can now move char_ffi to be in kernel/ so we dont need to duplicate the same code for every new character driver.