Read and Write return the amount of bytes read/written +logic fix
The read and write Uio traits now return a usize of the amount of bytes read/written. I also had fixed my poor write logic as I believe I wasn't treating the amount of bytes properly. I now need to fix the case for a null read/write operation as it seems to fail that test.